Add files via upload
multi-cross-modality one class cross learning approach.
Positive Only Learning is a machine learning approach that focuses solely on the positive instances of a dataset while disregarding negative examples. This method is particularly useful in scenarios where negative samples are scarce or difficult to obtain.
Negative Only Learning is a complementary approach where the model is trained exclusively on negative examples without any positive instances. This method is beneficial in situations where positive samples are limited or hard to define.